Remove @statechannels/wallet dependency by hub
