+++ packages/hub/package.json
@@ -33,7 +33,6 @@
   "dependencies": {
     "@koa/cors": "2.2.3",
     "@statechannels/nitro-protocol": "0.0.1",
-    "@statechannels/wallet": "^0.0.1",
     "async-lock": "1.2.2",
     "dotenv": "6.2.0",
     "dotenv-expand": "4.2.0",
