+++ packages/hub/src/hub/handlers/handle-new-process-action.ts
@@ -1,10 +1,9 @@
-import {unreachable} from '@statechannels/wallet';
 import {
   ConcludeInstigated,
-  SignedStatesReceived,
-  RelayActionWithMessage
+  RelayActionWithMessage,
+  SignedStatesReceived
 } from '@statechannels/wallet/lib/src/communication';
-import {HUB_ADDRESS} from '../../constants';
+import {HUB_ADDRESS, unreachable} from '../../constants';
 import {startFundingProcess} from '../../wallet/db/queries/walletProcess';
 import * as ongoing from './handle-ongoing-process-action';
