+++ packages/hub/src/wallet/utilities/hex-utils.ts
@@ -1,0 +1,7 @@
+import {bigNumberify} from 'ethers/utils';
+
+export function addHex(a: string, b: string): string {
+  return bigNumberify(a)
+    .add(bigNumberify(b))
+    .toHexString();
+}
